LPS-52265 move model listeners to module
