+++ portal-impl/src/META-INF/model-listener-spring.xml
@@ -8,12 +8,9 @@
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
 >
 	<bean id="com.liferay.portal.model.BackgroundTaskModelListener" class="com.liferay.portal.model.BackgroundTaskModelListener" />
-	<bean id="com.liferay.portal.model.ContactModelListener" class="com.liferay.portal.model.ContactModelListener" />
 	<bean id="com.liferay.portal.model.LayoutModelListener" class="com.liferay.portal.model.LayoutModelListener" />
 	<bean id="com.liferay.portal.model.LayoutSetModelListener" class="com.liferay.portal.model.LayoutSetModelListener" />
 	<bean id="com.liferay.portal.model.PortletPreferencesModelListener" class="com.liferay.portal.model.PortletPreferencesModelListener" />
-	<bean id="com.liferay.portal.model.UserGroupModelListener" class="com.liferay.portal.model.UserGroupModelListener" />
-	<bean id="com.liferay.portal.model.UserModelListener" class="com.liferay.portal.model.UserModelListener" />
 	<bean id="com.liferay.portal.service.impl.LayoutSetPrototypeLayoutModelListener" class="com.liferay.portal.service.impl.LayoutSetPrototypeLayoutModelListener" />
 	<bean id="com.liferay.portal.service.impl.LayoutSetPrototypeLayoutSetModelListener" class="com.liferay.portal.service.impl.LayoutSetPrototypeLayoutSetModelListener" />
 	<bean id="com.liferay.portlet.asset.model.AssetEntryModelListener" class="com.liferay.portlet.asset.model.AssetEntryModelListener" />
